ci/run: Extend and refactor run script with '-p(lan)' option
Extend the run script with an additional '-p' option to be able to run a
specific plan and refactor the script to support the latest nci-config
plans with deploy modes automatically selecting default deploy modes for
all bob recipes (i.e. sdcard for qemu, tftp for xilinx and log for all
prove related recipes). In addition, the error handling is adjusted for
the bob build command with 'exit 1' if the build of one of the recipes
fails.