muen_component: Remove -m from gprinstall
Option -a is the default (copy all sources). This is beneficial for
gnatmetric since we also want to see *all* bodies of libs when running
gnatmetric in the kernel package.

Note that the Externally_Built attribute is still set to True by
gprinstall, "indicating to the builder that this project can be assumed
to be up-to-date, and should not be considered for recompilation." [1]

[1] - https://docs.adacore.com/gprbuild-docs/html/gprbuild_ug/gnat_project_manager.html#importing-projects