Fix mucheckstack in bob build
Summary:
Closes T770. This includes changes to generate .ci files in common
libraries and collect them in the corresponding -dev packages.
mucheckstack is changed to find the files in that location next to
the GPR files. Finally, the OBJ_DIR environment variable is set when
running mucheckstack so the .ci files of the checked application are
found in the build tree.

Reviewers: #muen, reet

Reviewed By: #muen, reet

Subscribers: reet

Maniphest Tasks: T770

Differential Revision: https://dev.codelabs.ch/D1252